+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes in Element 1.0.9 (2020-XX-XX)
 Features ✨:
  - Search messages in a room - phase 1 (#2110)
  - Hide encrypted history (before user is invited). Can be shown if wanted in developer settings
+ - Changed rainbow algorithm
 
 Improvements 🙌:
  - Wording differentiation for direct rooms (#2176)